fixed ata_write_master_no_no_ow

+++ arch/x86/fs/WIP/inffs/inffs.h
@@ -62,5 +62,5 @@ struct inffs_path{
 struct __INFFS_FILE *__INFFS_FULLDISK_FS_FOPEN(const char *path,int opperation,struct __INFFS_FILE *ret);
 int __INFFS_MKFS_FULLDISK();
 int __INFFS_FULLDISK_FS_FREAD(struct __INFFS_FILE *file,int *buf,int n);
-int __INFFS_FULLDISK_FS_FWRITE(struct __INFFS_FILE *file,int *buf,int n);
+int __INFFS_FULLDISK_FS_FWRITE(struct __INFFS_FILE *file,uint8_t *buf,int n);
 #endif /* defined(____inffs__) */
